Add summary field to user onboarding details
The code changes include adding a summary field to the user onboarding details. This allows users to provide a summary of their onboarding process. The summary field has been added to the state object and is included in the request body when sending the user details to the server using the API endpoint.